Firewall error handling and security fixes
The Shell class now provides the developer with an error if the command wrote something to stderr.\nI added this to the firewall new rule and delete rule functionality. This prevents the firewall from getting out of hand. Instead of having to write a guard clause for every API functionality, I am now using a middelwear that checks for admin priveleges.\nI also added some protections against DoS attacks.