Added tasks for slow requests
I added background tasks to Zentrox. When a task is expected to take longer than most browsers fetch timeout,
a task is used. This means that the backend will start a seperat thread running the task and will return a UUID to the frontend.
The frontend can now fetch the tasks status periodically to check if the task is working, finished or failed.
In addition, I removed irrelevant println! calls that were used to check how long it takes for a dashboard value to be determined.