Improved Overview loading speed
Instead of fetching the packageDatabase information before the
deviceInformation, the Overview will now switch those two fetches. This
way the frontend will wait for the deviceInformation to be loaded and
only then will it ask for the packageDatabase. By doing this, the
majority of the content is loaded before the minority and you will not
have a very slow task be running on the backend at the same time as a
very fast task.