Replaced database.rs library with diesel.rs and added file sharing backend
I have largely removed the SQL database connection layer using rusqlite
which operated through direct SQL queries and no paramertization with a
backend powered using diesel.rs.
This means, that I removed all database query shorthands that rely on
simple SQL queries and connections to the database file will mostly be
established using the establish_connection() function in database.rs.
The database.rs only contains a few shorthands for reading values from
the key-value databases Secrets and Settings as well as the setup
function that creates the SQL tables using the setup.sql code.
Ideally, this will also all be re-written to use diesel. Key-Value
tables should also be replaced with better alternatives, as storing a
value of known type with a known key as a string just to fit in with the
Text-Text database schema is chaotic and painful.
The code now relies on the models.rs and schema.rs files to provide
models and database schemas for diesel.rs to work with.

Lastly, I also added the backend for file sharing using a key and
password protection. This feature has not been thoroughly tested.