Fixed bug not sending username on page reload
Some browsers may cache the value set in the username input, which does
not trigger onChange for the input field, thus not updating the state
that stores the password.
This was fixed by using a ref.current.value for both username and
password input to read those values.
I also addded a nice fade out animation.