Added file sharing frontend, rate-limiting and metadata endpoint
Zentrox now includes a frontend for sharing, unsharing and downloading
files with and without passwords. The sharing and unsharing interface is
integrated in the file manager on the admin protected side of the
application. The backend now also keeps track of the time the sharing
was created (relative to UNIX Epoch) and transfers this information to
the frontend)  The frontend for downloading files is located under
/shared and takes the get parameter code. When accessed, the page
requests metadata about the code using /shared/getMetadata.  The file
can then be downloaded from /shared/get. A password may have to be
provided.
All three endpoints under /shared are rate limited to 9 requests per
minute. For this, a new governor (shared_file_governor) was implemented.