Rust directory structure prepaired and experimented
I have setup the basic directory structure for having a library and a
binary crate as well as modules for organizing smaller helpers and
routes. Everything has been re-wired to work, content has not been
filled into the structure.
This is a valuable to towards separating the different components of the
code.