Fixed bug with number for time and improved library
I fixed a bug that caused the library to panic when parsing journalctl
entries. If the entries used a number to store the date instead of a
string, the program was not prepared and would fail.
I have replaced basic u64 for communicating time with the library with
Duration and rewrote small parts of code to be more idiomatic.